requirements.txt
Django==4.2
djangorestframework==3.14.0
python-dotenv==1.0.0
Pillow==10.3.0
django-crispy-forms==1.14.0
django-filter==23.2.0